refactor(#6): CheckUserPort 통합 - CheckUserEmailPort와 CheckUserPhoneNumberPort 모두 유저의 존재여부를 확인하는 메서드이므로 CheckUserPort로 합쳤어요.
