+++ src/main/java/com/teka/application/chatroom/service/RegisterUserService.java
@@ -1,21 +1,18 @@
 package com.teka.application.chatroom.service;
 
 import com.teka.application.chatroom.port.out.FindChatRoomPort;
-import com.teka.domain.chatroom.exception.ChatRoomAdminMismatchException;
 import com.teka.application.chatroom.exception.ChatRoomNotFoundException;
 import com.teka.application.chatroom.port.in.RegisterUserUseCase;
 import com.teka.application.chatroom.port.in.command.RegisterUserCommand;
-import com.teka.application.user.port.in.exception.EmailAlreadyExistsException;
-import com.teka.application.user.port.in.exception.PhoneNumberAlreadyExistsException;
-import com.teka.application.user.port.out.CheckUserEmailPort;
-import com.teka.application.user.port.out.CheckUserPhoneNumberPort;
+import com.teka.application.user.exception.EmailAlreadyExistsException;
+import com.teka.application.user.exception.PhoneNumberAlreadyExistsException;
+import com.teka.application.user.port.out.CheckUserPort;
 import com.teka.application.user.port.out.SaveUserPort;
 import com.teka.domain.admin.AdminId;
 import com.teka.domain.chatroom.ChatRoomId;
 import com.teka.domain.user.Email;
 import com.teka.domain.user.PhoneNumber;
 import com.teka.domain.user.User;
-import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -26,8 +23,7 @@ import java.util.List;
 public class RegisterUserService implements RegisterUserUseCase {
 
     private final SaveUserPort saveUserPort;
-    private final CheckUserPhoneNumberPort checkUserPhoneNumberPort;
-    private final CheckUserEmailPort checkUserEmailPort;
+    private final CheckUserPort checkUserPort;
     private final FindChatRoomPort findChatRoomPort;
 
     @Override
@@ -36,10 +32,10 @@ public class RegisterUserService implements RegisterUserUseCase {
                 .orElseThrow(ChatRoomNotFoundException::new)
                 .isAdmin(adminId);
         for (RegisterUserCommand command : commandList) {
-            if (checkUserPhoneNumberPort.existsByPhoneNumber(command.phoneNumber(), chatRoomId)) {
+            if (checkUserPort.existsByPhoneNumber(new PhoneNumber(command.phoneNumber()), chatRoomId)) {
                 throw new PhoneNumberAlreadyExistsException();
             }
-            if (checkUserEmailPort.existsByEmail(command.email(), chatRoomId)) {
+            if (checkUserPort.existsByEmail(command.email(), chatRoomId)) {
                 throw new EmailAlreadyExistsException();
             }
